Move roll context to separate file
